+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+* 0.1.7
+    * ssh log monitoring
+    * rules for text now support "is" and "is not" conditions
+
 * 0.1.6
     * safe numbers conversions
     * ids collision tests